Silence warning in old gcc

+++ src/Monkeypatcher.cc
@@ -778,7 +778,7 @@ template <> void patch_after_exec_arch<X64Arch>(Task* t, Monkeypatcher&) {
         // The symbol values can be absolute or relative addresses.
         // The first part of the assertion is for absolute
         // addresses, and the second part is for relative.
-        ASSERT(t, (sym_address & ~vdso_max_size) == vdso_static_base ||
+        ASSERT(t, uint64_t(sym_address & ~vdso_max_size) == vdso_static_base ||
                       (sym_address & ~vdso_max_size) == 0);
         uintptr_t sym_offset = sym_address & vdso_max_size;
         uintptr_t absolute_address = vdso_start.as_int() + sym_offset;
